Omflyttning av funktioner till fler filer
* Datatyper som alla filer använder hamnade i datatypes.h
* Tanken är att man sen ska kunna inkludera typ bara motion.h
